backend/pyproject.toml
[project]
name = "pricepilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"fastapi>=0.116.0",
	"hijridate>=2.5.0",
	"holidays>=0.71",
	"numpy>=2.3.0",
	"pandas>=2.3.0",
	"python-dotenv>=1.1.0",
	"python-dateutil>=2.9.0.post0",
	"requests>=2.32.0",
	"scikit-learn>=1.7.0",
	"uvicorn>=0.35.0",
]
